Refactor player components
